weather_agent/requirements.txt
langchain
langchain-openai
python-dotenv
serpapi
requests